requirements.txt
numpy>=1.23
tqdm>=4.66
tabulate>=0.9

openai>=1.12

torch>=2.0
transformers>=4.41
accelerate>=0.28

evaluate>=0.4
datasets>=2.14
pandas>=2.0
google-generativeai>=0.5
rouge-score>=0.1.2
sentencepiece>=0.1.99
